Refactor css and file structure Convert font sizes to rem with media queries, Move image styling to css class, Replace 100vh/vw wrapper size with fixed position, Center wrapper vertically
